Use wider read_small

+++ src/operations.rs
@@ -16,19 +16,19 @@ pub(crate) const fn folded_multiply(s: u64, by: u64) -> u64 {
 /// Given a small (less than 8 byte slice) returns the same data stored in two u32s.
 /// (order of and non-duplication of bytes is NOT guaranteed)
 #[inline(always)]
-pub(crate) fn read_small(data: &[u8]) -> [u32; 2] {
+pub(crate) fn read_small(data: &[u8]) -> [u64; 2] {
     debug_assert!(data.len() <= 8);
     if data.len() >= 2 {
         if data.len() >= 4 {
             //len 4-8
-            [data.read_u32().0 as u32, data.read_last_u32() as u32]
+            [data.read_u32().0 as u64, data.read_last_u32() as u64]
         } else {
             //len 2-3
-            [data.read_u16().0 as u32, data[data.len() - 1] as u32]
+            [data.read_u16().0 as u64, data[data.len() - 1] as u64]
         }
     } else {
         if data.len() > 0 {
-            [data[0] as u32, 0]
+            [data[0] as u64, 0]
         } else {
             [0, 0]
         }
